pyproject.toml
[tool.poetry]
name = "gaphas"
version = "5.0.0"
description="Gaphas is a GTK diagramming widget"
authors = [
    "Arjan J. Molenaar <gaphor@gmail.com>",
    "Dan Yeaw <dan@yeaw.me>"
]
license = "Apache-2.0"

readme = "README.md"

homepage = "https://gaphas.readthedocs.io/"
repository = "https://github.com/gaphor/gaphas"
documentation = "https://gaphas.readthedocs.io/"

keywords = ["gtk", "diagram", "gaphas"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.9,<4"
PyGObject = "^3.50"
pycairo = "^1.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^5.0"
pytest-archon = "^0.0.6"
pytest-asyncio = ">=0.23.8,<0.25.0"

[tool.poetry.group.docs]
optional=true

[tool.poetry.group.docs.dependencies]
sphinx = ">=4.3,<8.0"
furo = ">=2022,<2025"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.coverage.run]
source = ["gaphas"]

[tool.mypy]
python_version = 3.9
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
check_untyped_defs = true
strict_optional = true
disallow_any_explicit = true
show_error_codes = true
ignore_missing_imports=true
namespace_packages = true
explicit_package_bases = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "cairo",
    "gi.*",
    "hotshot.*",
]
ignore_missing_imports = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "gaphas.*"
disallow_incomplete_defs = true

[tool.ruff]
exclude = [
    ".venv",
    "dist",
    "__init__.py",
]
line-length = 88
[tool.ruff.lint]
ignore = ["E501", "B905", "B019"]
select = [
    "B",
    "B9",
    "C",
    "E",
    "F",
    "W",
]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.codespell]
skip = 'README.md, CODE_OF_CONDUCT.md'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

